Unify code. Only get coinbase data in tiingo rest request. fix join hanging by specifying timeout.
